fix goldilock test build

+++ field/goldilocks/extensions/e2_test.go
@@ -523,8 +523,9 @@ func genFr() gopter.Gen {
 		// to have a deterministic generator.
 		e := bigIntPool.Get().(*big.Int)
 		e.Rand(genParams.Rng, modulus)
+
 		for i, w := range e.Bits() {
-			elmt[i] = uint32(w)
+			elmt[i] = uint64(w)
 		}
 		bigIntPool.Put(e)
